F [#CSS-40] Renamed gem to shore-customers
