+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in customer_store-client.gemspec
+# Specify your gem's dependencies in shore-customers.gemspec
 gemspec
