--- lib/customer_store/client/version.rb
@@ -1,5 +1,0 @@
-module CustomerStore
-  class Client # :nodoc:
-    VERSION = '0.1.0'
-  end
-end
