+++ lib/shore/customers.rb
@@ -1,0 +1,2 @@
+require 'shore/customers/version'
+require 'shore/customers/client'
