+++ lib/shore/customers/version.rb
@@ -1,0 +1,5 @@
+module Shore
+  module Customers # :nodoc:
+    VERSION = '0.0.2'
+  end
+end
