+++ spec/support/customer_store.rb
@@ -1,10 +1,10 @@
 require_relative 'webmock'
-require 'customer_store/rspec'
+require 'shore/customers/rspec'
 
 RSpec.configure do |config|
   config.before(:each) do
-    ENV['CUSTOMER_STORE_BASE_URI'] = CustomerStore::Rspec::DEFAULT_BASE_URI
+    ENV['CUSTOMER_STORE_BASE_URI'] = Shore::Customers::Rspec::DEFAULT_BASE_URI
     ENV['CUSTOMER_STORE_SECRET'] = nil
-    config.include CustomerStore::Rspec
+    config.include Shore::Customers::Rspec
   end
 end
